# public/swagger-specs/services/transaction-scheduler/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "PegaSys Orchestrate Transaction API. Enables dynamic management of transactions.\nTransaction requests are an abstraction over schedules and jobs representing one or more transactions executed on the Blockchain network\nSchedules are ordered lists of jobs executed in a predefined sequence\nJobs represent a single transaction flow executed on the Blockchain network",
        "title": "Transaction Scheduler API",
        "contact": {
            "name": "Contact PegaSys Orchestrate",
            "url": "https://pegasys.tech/contact/",
            "email": "support@pegasys.tech"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0"
    },
    "paths": {
...
        "/schedules": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "JWTAuth": []
                    }
                ],
                "description": "Get all schedules",
                "produces": [
                    "application/json"
                ],
                "summary": "Get all schedules",
                "responses": {
                    "200": {
                        "description": "List of schedules found",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/txscheduler.ScheduleResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "JWTAuth": []
                    }
                ],
                "description": "Creates a new schedule",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Creates a new Schedule",
                "parameters": [
                    {
                        "description": "Schedule creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/txscheduler.CreateScheduleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Created schedule",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/txscheduler.ScheduleResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "jobs": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/txscheduler.JobResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable parameters were sent",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/schedules/{uuid}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "JWTAuth": []
                    }
                ],
                "description": "Fetch a single schedule by uuid",
                "produces": [
                    "application/json"
                ],
                "summary": "Fetch a schedule by uuid",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID of the schedule",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Schedule found",
                        "schema": {
                            "$ref": "#/definitions/txscheduler.ScheduleResponse"
                        }
                    },
                    "404": {
                        "description": "Schedule not found",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "entities.ETHTransaction": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string",
                    "example": "2020-07-09T12:35:42.115395Z"
                },
                "data": {
                    "type": "string",
                    "example": "0xfe378324abcde723"
                },
                "enclaveKey": {
                    "type": "string",
                    "example": "0xd41551c714c8ec769d2edad9adc250ae955d263da161bf59142b7500eea6715eadc250ae955d263da161bf59142b7500eea6715e"
                },
                "from": {
                    "type": "string",
                    "example": "0x1abae27a0cbfb02945720425d3b80c7e09728534"
                },
                "gas": {
                    "type": "string",
                    "example": "21000"
                },
                "gasPrice": {
                    "type": "string",
                    "example": "71500000 (wei)"
                },
                "hash": {
                    "type": "string",
                    "example": "0xd41551c714c8ec769d2edad9adc250ae955d263da161bf59142b7500eea6715e"
                },
                "nonce": {
                    "type": "string",
                    "example": "1"
                },
                "privacyGroupID": {
                    "type": "string",
                    "example": "A1aVtMxLCUHmBVHXoZzzBgPbW/wj5axDpW9X8l91SGo="
                },
                "privateFor": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "A1aVtMxLCUHmBVHXoZzzBgPbW/wj5axDpW9X8l91SGo=",
                        "B1aVtMxLCUHmBVHXoZzzBgPbW/wj5axDpW9X8l91SGo="
                    ]
                },
                "privateFrom": {
                    "type": "string",
                    "example": "A1aVtMxLCUHmBVHXoZzzBgPbW/wj5axDpW9X8l91SGo="
                },
                "raw": {
                    "type": "string",
                    "example": "0xfe378324abcde723"
                },
                "to": {
                    "type": "string",
                    "example": "0x1abae27a0cbfb02945720425d3b80c7e09728534"
                },
                "updatedAt": {
                    "type": "string",
                    "example": "2020-07-09T12:35:42.115395Z"
                },
                "value": {
                    "type": "string",
                    "example": "71500000 (wei)"
                }
            }
        },
        "entities.Log": {
            "type": "object",
            "properties": {
                "at": {
                    "type": "string",
                    "example": "2020-07-09T12:35:42.115395Z"
                },
                "message": {
                    "type": "string",
                    "example": "Log message"
                },
                "status": {
                    "type": "string",
                    "example": "MINED"
                }
            }
        },
        "httputil.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 24000
                },
                "message": {
                    "type": "string",
                    "example": "invalid status update for the current job state"
                }
            }
        },
        "txscheduler.Annotations": {
            "type": "object",
            "properties": {
                "gasPricePolicy": {
                    "type": "object",
                    "$ref": "#/definitions/txscheduler.GasPriceParams"
                },
                "oneTimeKey": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "txscheduler.CreateScheduleRequest": {
            "type": "object"
        },
...
        "txscheduler.GasPriceParams": {
            "type": "object",
            "properties": {
                "priority": {
                    "type": "string",
                    "example": "very-high"
                },
                "retryPolicy": {
                    "type": "object",
                    "$ref": "#/definitions/txscheduler.RetryParams"
                }
            }
        },
...
        "txscheduler.JobResponse": {
            "type": "object",
            "properties": {
                "annotations": {
                    "type": "object",
                    "$ref": "#/definitions/txscheduler.Annotations"
                },
                "chainUUID": {
                    "type": "string",
                    "example": "b4374e6f-b28a-4bad-b4fe-bda36eaf849c"
                },
                "createdAt": {
                    "type": "string",
                    "example": "2020-07-09T12:35:42.115395Z"
                },
                "labels": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "logs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.Log"
                    }
                },
                "nextJobUUID": {
                    "type": "string",
                    "example": "b4374e6f-b28a-4bad-b4fe-bda36eaf849c"
                },
                "parentJobUUID": {
                    "type": "string",
                    "example": "b4374e6f-b28a-4bad-b4fe-bda36eaf849c"
                },
                "scheduleUUID": {
                    "type": "string",
                    "example": "b4374e6f-b28a-4bad-b4fe-bda36eaf849c"
                },
                "status": {
                    "type": "string",
                    "example": "MINED"
                },
                "transaction": {
                    "type": "object",
                    "$ref": "#/definitions/entities.ETHTransaction"
                },
                "type": {
                    "type": "string",
                    "example": "eth://ethereum/transaction"
                },
                "updatedAt": {
                    "type": "string",
                    "example": "2020-07-09T12:35:42.115395Z"
                },
                "uuid": {
                    "type": "string",
                    "example": "b4374e6f-b28a-4bad-b4fe-bda36eaf849c"
                }
            }
        },
...
        "txscheduler.RetryParams": {
            "type": "object",
            "properties": {
                "increment": {
                    "type": "number",
                    "example": 0.05
                },
                "interval": {
                    "type": "string",
                    "example": "2m"
                },
                "limit": {
                    "type": "number",
                    "example": 0.5
                }
            }
        },
        "txscheduler.ScheduleResponse": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string",
                    "example": "2020-07-09T12:35:42.115395Z"
                },
                "jobs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/txscheduler.JobResponse"
                    }
                },
                "tenantID": {
                    "type": "string",
                    "example": "tenant_id"
                },
                "uuid": {
                    "type": "string",
                    "example": "b4374e6f-b28a-4bad-b4fe-bda36eaf849c"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "JWTAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}